requirements.txt
streamlit
python-dotenv
langchain-core
langchain-groq
groq